Use Falcon for sinatra http server
