+++ Gemfile
@@ -4,6 +4,7 @@ ruby '>= 3.0.0'
 
 gem 'async'
 gem 'sinatra'
+gem 'falcon'
 gem 'puma'
 gem 'async-http'
 gem "httpray"
